refactored frontend components into new folders
