+++ frontend/src/components/competition/CompetitionResult/UserResult.jsx
@@ -1,5 +1,5 @@
 import {useState, useEffect} from 'react'
-import styles from "../../styles/Questions.module.css"
+import styles from "../../../styles/Questions.module.css"
 
 const UserResult = ({attempt,questions,amountRight}) => {
