+++ frontend/src/components/competition/TableDisplay/CompetitionDetailsParramTitles.jsx
@@ -1,4 +1,4 @@
-import styles from '../../styles/Landing.module.css'
+import styles from '../../../styles/Landing.module.css'
 
 
 const CompetitionDetailsParramTitles = () => {
