+++ frontend/src/components/competition/popups/CompetitionDetailsPopup.jsx
@@ -1,12 +1,12 @@
-import style from '../../styles/UserPopup.module.css';
+import style from '../../../styles/UserPopup.module.css';
 import {useState,useEffect} from 'react';
-import Dropdown from '../misc/Dropdown.jsx';
-import eventService from '../../services/eventService.js';
-import userService from '../../services/userService.js';
-import QuestionDisplay from '../question/QuestionDisplay.jsx';
-import CompetitionFillout from './CompetitionFillout.jsx';
-import competitionService from '../../services/competitionService.js';
-import EventDropDownList from '../event/EventDropDownList.jsx';
+import Dropdown from '../../misc/Dropdown.jsx';
+import eventService from '../../../services/eventService.js';
+import userService from '../../../services/userService.js';
+import QuestionDisplay from '../../question/TableDisplay/QuestionDisplay.jsx';
+import CompetitionFillout from '../UserInput/CompetitionFillout.jsx';
+import competitionService from '../../../services/competitionService.js';
+import EventDropDownList from '../../event/EventDropDownList.jsx';
 
 const CompetitionDetailsPopup = ({setViewPopupOpen, competition,hasJoinedEvent}) => {
     const [questionAmount, setQuestionAmount] = useState(0);
