+++ frontend/src/components/event/TableDisplay/EventDisplay.jsx
@@ -1,13 +1,13 @@
 
-import styles from '../../styles/Landing.module.css'
+import styles from '../../../styles/Landing.module.css'
 import { useEffect, useState } from 'react';
 import EventTableEntry from './EventTableEntry.jsx'
 
-import userService from '../../services/userService.js';
-import eventService from '../../services/eventService.js';
+import userService from '../../../services/userService.js';
+import eventService from '../../../services/eventService.js';
 import EventDetailsParramTitles from './EventDetailsParramTitles.jsx';
-import Dropdown from '../misc/Dropdown.jsx';
-import CreateEvent from './CreateEvent.jsx';
+import Dropdown from '../../misc/Dropdown.jsx';
+import CreateEvent from '../CreateEvent.jsx';
 
 
 const EventDisplay = ({ user }) => {
