+++ frontend/src/components/question/TableDisplay/QuestionTableEntry.jsx
@@ -1,4 +1,4 @@
-import styles from '../../styles/Questions.module.css'
+import styles from '../../../styles/Questions.module.css'
 //displays the question in a table format
 const QuestionTableEntry = ({ question }) => {
     return (
